+++ be/src/exec/pipeline/nljoin/nljoin_context.h
@@ -56,7 +56,7 @@ public:
 
     int32_t get_num_builders() const { return _num_right_sinkers; }
     bool is_build_chunk_empty() const { return _build_chunks.empty(); }
-    int32_t num_build_chunks() const { return _build_chunks.size(); }
+    size_t num_build_chunks() const { return _build_chunks.size(); }
     size_t num_build_rows() const { return _num_build_rows; }
 
     Chunk* get_build_chunk(int32_t index) const { return _build_chunks[index].get(); }
